// src/interfaces/refs/refs.ts
import { MutableRefObject } from "react";
export interface IPassRefs {
  home: MutableRefObject<null>;
  experiences: MutableRefObject<null>;
  techniques: MutableRefObject<null>;
  process: MutableRefObject<null>;
  play: MutableRefObject<null>;
  projects: MutableRefObject<null>;
  code: MutableRefObject<null>;
  about: MutableRefObject<null>;
  connect: MutableRefObject<null>;
}